Pull expectedNpmVersionFailures from GitHub, like allowedPackageJsonDependencies
