+++ .changeset/moody-tomatoes-promise.md
@@ -1,0 +1,7 @@
+---
+"@definitelytyped/definitions-parser": patch
+"@definitelytyped/dtslint": patch
+"@definitelytyped/utils": patch
+---
+
+Pull expectedNpmVersionFailures from GitHub, like allowedPackageJsonDependencies
