+++ .github/workflows/ci.yml
@@ -62,5 +62,5 @@ jobs:
         fi
 
         git switch -c changesets-temp
-        git checkout origin/main -- packages/definitions-parser/allowedPackageJsonDependencies.txt
+        git checkout origin/main -- packages/definitions-parser/allowedPackageJsonDependencies.txt packages/dtslint/expectedNpmVersionFailures.txt
         pnpm changeset status --since=origin/main
