+++ packages/utils/src/io.ts
@@ -300,3 +300,17 @@ function addExecutePermissionsFromReadPermissions(mode: number): number {
   // Add these additional execute permissions to the mode.
   return mode | readPermissionsAsExecutePermissions;
 }
+
+export function getUrlContentsAsString(url: string): Promise<string> {
+  return new Promise<string>((resolve, reject) => {
+    https
+      .get(url, (res) => {
+        let data = "";
+        res.on("data", (d) => (data += d));
+        res.on("end", () => {
+          resolve(data);
+        });
+      })
+      .on("error", reject);
+  });
+}
